Fixes type conversion warning

+++ src/map/status.cpp
@@ -8730,7 +8730,7 @@ t_tick status_get_sc_def(struct block_list *src, struct block_list *bl, enum sc_
 		sc = NULL;
 
 #ifdef RENEWAL
-	uint16 levelAdv = (pow(max(0, status_get_lv(src) - status_get_lv(bl)), 2) / 5) * 100;
+	uint16 levelAdv = static_cast<uint16>(pow(max(0, status_get_lv(src) - status_get_lv(bl)), 2) / 5) * 100;
 #endif
 
 	switch (type) {
